Correct postprocess error "out of bound" due to a grep that retrieves rows of files such "14_xx.fcs" and "4_xx.fcs" when looking for the second second one. Correct shinyApp error while removing heatmap.

+++ R/cytofkit_shinyAPP.R
@@ -1184,7 +1184,8 @@ cytofkitShinyAPP <- function(RData = NULL, onServer = FALSE) {
           })
           
           session$onSessionEnded(function(){
-            file.remove("cytofkit_shinyAPP_marker_heatmap.pdf")
+            if (file.exists("cytofkit_shinyAPP_marker_heatmap.pdf"))
+              file.remove("cytofkit_shinyAPP_marker_heatmap.pdf")
           })
           
           ##-----level plot-----
